requirements.txt
httpx>=0.27.0
pyyaml>=6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
exa-py>=1.0.0
